rename: 변경된 asset 폴더구조 변경
